fix: servo controller
