+++ main2.js
@@ -35,13 +35,9 @@ mqttClient.on('connect', function () {
   })
 })
 
-const { Board, Fn } = require('johnny-five')
-const servoType = require('./src/servoType')
-
+const { Board } = require('johnny-five')
+const servoController = require('./src2/servoController')
 const board = new Board({ port: ARDUINO_PATH, repl: false })
-
-const bodyServo = new servoType()
-const cal = (x) => Fn.map(x, 0, 320, 0, 288)
 
 const TB6612_AIN1 = 0
 const TB6612_AIN2 = 1
@@ -58,55 +54,6 @@ board.on('ready', () => {
   board.digitalWrite(TB6612_STBY, 1)
 
   mqttClient.on('message', function (topic, message) {
-    if (topic === `${email}/command`) {
-      bodyServo.neckPitch.to(cal(message[3]))
-      bodyServo.neckYaw.to(cal(message[4]))
-      bodyServo.leftArmPitch.to(cal(message[5]))
-      bodyServo.leftArmYaw.to(cal(message[6]))
-      bodyServo.rightArmPitch.to(cal(message[7]))
-      bodyServo.rightArmYaw.to(cal(message[8]))
-
-      if (message[9] === 50 && message[10] === 50) {
-        bodyServo.leftMotor.stop()
-        bodyServo.rightMotor.stop()
-      }
-      if (message[9] === 100 && message[10] === 100) {
-        board.digitalWrite(TB6612_AIN1, 1)
-        board.digitalWrite(TB6612_AIN2, 0)
-        bodyServo.leftMotor.speed(150)
-        board.digitalWrite(TB6612_BIN1, 0)
-        board.digitalWrite(TB6612_BIN2, 1)
-        bodyServo.rightMotor.speed(150)
-      }
-      if (message[9] === 0 && message[10] === 0) {
-        board.digitalWrite(TB6612_AIN1, 0)
-        board.digitalWrite(TB6612_AIN2, 1)
-        bodyServo.leftMotor.speed(150)
-
-        board.digitalWrite(TB6612_BIN1, 1)
-        board.digitalWrite(TB6612_BIN2, 0)
-        bodyServo.rightMotor.speed(150)
-      }
-
-      if (message[9] === 100 && message[10] === 0) {
-        board.digitalWrite(TB6612_AIN1, 0)
-        board.digitalWrite(TB6612_AIN2, 1)
-        bodyServo.leftMotor.speed(150)
-
-        board.digitalWrite(TB6612_BIN1, 0)
-        board.digitalWrite(TB6612_BIN2, 1)
-        bodyServo.rightMotor.speed(150)
-      }
-
-      if (message[9] === 0 && message[10] === 100) {
-        board.digitalWrite(TB6612_AIN1, 1)
-        board.digitalWrite(TB6612_AIN2, 0)
-        bodyServo.leftMotor.speed(150)
-
-        board.digitalWrite(TB6612_BIN1, 1)
-        board.digitalWrite(TB6612_BIN2, 0)
-        bodyServo.rightMotor.speed(150)
-      }
-    }
+    servoController(board, topic, message)
   })
 })
